--- front/app/containers/Admin/projects/project/insights/config/sectionConfig.ts
@@ -1,8 +1,0 @@
-import { ParticipationMethod } from 'api/phases/types';
-
-/**
- * Check if a method is idea-based (ideation or proposals).
- */
-export function isIdeaBasedMethod(method?: ParticipationMethod): boolean {
-  return method === 'ideation' || method === 'proposals';
-}
