--- front/app/containers/Admin/projects/project/insights/word/insightConverters/types.ts
@@ -1,14 +1,0 @@
-import { MessageDescriptor } from 'react-intl';
-
-import type { Localize } from 'hooks/useLocalize';
-
-import { FormatMessageValues } from 'utils/cl-intl/useIntl';
-
-export interface WordExportIntl {
-  formatMessage: (
-    descriptor: MessageDescriptor,
-    values?: FormatMessageValues
-  ) => string;
-  locale: string;
-  localize: Localize;
-}
